Create Task_order_array_recurr.c
There are a number of task types determined by the user, each type of task will appear with a given probability at the beginning of each hour, for the number of hours determined by the user. The tasks are ordered with importance that is scaled by the penalty function, which is related to the completion time.